# docs/swagger.json
{"swagger":"2.0","info":{"contact":{}},"paths":{"/index":{"get":{"consumes":["application/json"],"tags":["首页"],"responses":{"200":{"description":"OK","schema":{"type":"string"}}}}},"/user/createUser":{"post":{"consumes":["multipart/form-data"],"produces":["application/json"],"tags":["用户模块"],"summary":"新增用户","parameters":[{"type":"string","description":"用户名","name":"name","in":"formData","required":true},{"type":"string","description":"密码","name":"password","in":"formData","required":true},{"type":"string","description":"确认密码","name":"repassword","in":"formData","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":true}},"400":{"description":"Bad Request","schema":{"type":"object","additionalProperties":true}}}}},"/user/deleteUser":{"delete":{"produces":["application/json"],"tags":["用户模块"],"summary":"删除用户","parameters":[{"type":"integer","description":"用户ID","name":"id","in":"query","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":true}},"400":{"description":"Bad Request","schema":{"type":"object","additionalProperties":true}}}}},"/user/findUserByNameAndPwd":{"post":{"tags":["用户模块"],"summary":"所有用户","parameters":[{"type":"string","description":"用户名","name":"name","in":"query"},{"type":"string","description":"密码","name":"password","in":"query"}],"responses":{"200":{"description":"code\",\"message\"}","schema":{"type":"string"}}}}},"/user/list":{"get":{"description":"获取系统中所有用户的列表信息","produces":["application/json"],"tags":["用户模块"],"summary":"获取所有用户列表","responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":true}}}}},"/user/updateUser":{"put":{"consumes":["multipart/form-data"],"produces":["application/json"],"tags":["用户模块"],"summary":"修改用户信息","parameters":[{"type":"integer","description":"用户ID","name":"id","in":"formData","required":true},{"type":"string","description":"用户名","name":"name","in":"formData"},{"type":"string","description":"密码","name":"password","in":"formData"},{"type":"string","description":"电话","name":"phone","in":"formData"},{"type":"string","description":"邮箱","name":"email","in":"formData"}],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":true}},"400":{"description":"Bad Request","schema":{"type":"object","additionalProperties":true}}}}}}}